+++ Futures/Result.swift
@@ -8,31 +8,4 @@ public enum Result<T> {
   
   /// The result was cancelled
   case Cancelled
-  
-  private func _map<U>(handler: (T, Promise<U>) -> Void) -> Future<U> {
-    let mapped = Promise<U>()
-    
-    switch self {
-    case .Success(let value):
-      handler(value, mapped)
-    case .Error(let error):
-      mapped.fail(error)
-    case .Cancelled:
-      mapped.cancel()
-    }
-    
-    return mapped.future
-  }
-  
-  func map<U>(handler: T -> U) -> Future<U> {
-    return _map { value, mapped in
-      mapped.succeed(handler(value))
-    }
-  }
-  
-  func flatMap<U>(handler: T -> Future<U>) -> Future<U> {
-    return _map { value, flatMapped in
-      flatMapped.mimic(handler(value))
-    }
-  }
 }